+++ app/controllers/infrastructures_controller.rb
@@ -2,6 +2,10 @@
 
 class InfrastructuresController < ApplicationController
   def index
-    @infrastructures = FindDalmatianConfiguration.new.infrastructures
+    @infrastructures = Infrastructure.all
+  end
+
+  def show
+    @infrastructure = Infrastructure.find(params[:id])
   end
 end