+++ spec/services/find_dalmatian_configuration_spec.rb
@@ -1,19 +1,20 @@
 require "rails_helper"
 
 RSpec.describe FindDalmatianConfiguration do
-  describe "#infrastructures" do
-    it "returns the names of each infrastructure" do
+  describe "#call" do
+    it "returns a hash of the contents of dalmatian.yml" do
       fake_config = File.read("spec/fixtures/dalmatian-config/dalmatian.yml")
       allow(File).to receive(:read).and_return(fake_config)
 
-      result = described_class.new.infrastructures
+      result = described_class.new.call
 
-      expect(result).to match(
+      expect(result.keys).to match(
         [
-          "new-dedicated-cluster",
-          "shared-new-cluster",
-          "existing-shared-cluster-staging",
-          "existing-shared-cluster-production"
+          "parameter-store-path-prefix",
+          "account-bootstrap",
+          "ci",
+          "infrastructure-defaults",
+          "infrastructures"
         ]
       )
     end