+++ CHANGELOG.md
@@ -6,5 +6,8 @@ The format is based on [Keep a Changelog 1.0.0].
 
 ## [Unreleased]
 
+- all infrastructures are listed
+- individual infrastructures can be viewed
+
 [unreleased]: TODO
 [keep a changelog 1.0.0]: https://keepachangelog.com/en/1.0.0/